refac: rename interface to interfaces, chore: add response type
