+++ interfaces/list_wallets_response.go
@@ -1,4 +1,4 @@
-package _interface
+package interfaces
 
 import "github.com/DE-labtory/zulu/types"
