+++ server/api/walletapi.go
@@ -2,7 +2,6 @@ package api
 
 import (
 	"errors"
-
 	"github.com/DE-labtory/zulu/account"
 	"github.com/DE-labtory/zulu/keychain"
 	"github.com/DE-labtory/zulu/types"
